Day 52.java

+++ src/Library/Util.java
@@ -467,10 +467,10 @@ public class Util {
     }
 
 
-    public static void sleep(int seconds) {
+    public static void sleep(double seconds) {
 
         try {
-            Thread.sleep(seconds * 1000);
+            Thread.sleep((long)(seconds * 1000));
         } catch (InterruptedException e) {
 
         }
